refactor: logic to external file
